# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ByteBurrow API",
    "description": "REST API for ByteBurrow personal cloud storage",
    "contact": {
      "name": "Martin Miksanik"
    },
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/face/contacts": {
      "get": {
        "tags": [
          "face"
        ],
        "summary": "List contacts with their face counts\nGET /api/face/contacts",
        "operationId": "list_contacts_handler",
        "responses": {
          "200": {
            "description": "All contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContactResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "face"
        ],
        "summary": "Create a contact (a named person)\nPOST /api/face/contacts",
        "operationId": "create_contact_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Contact created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Contact name already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/face/contacts/{id}": {
      "put": {
        "tags": [
          "face"
        ],
        "summary": "Rename a contact\nPUT /api/face/contacts/:id",
        "operationId": "update_contact_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Contact ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Contact renamed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Contact name already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "face"
        ],
        "summary": "Delete a contact. Its face rows cascade away (FK), shrinking the exemplar\npool, so a full re-match pass is queued afterwards — remaining suggestions\nthat only matched this person get withdrawn.\nDELETE /api/face/contacts/:id",
        "operationId": "delete_contact_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Contact ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact deleted, re-match queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/face/refs": {
      "get": {
        "tags": [
          "face"
        ],
        "summary": "List detected faces (the review queue)\nGET /api/face/refs",
        "operationId": "list_face_refs_handler",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "1-based page number (default 1).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Items per page (default 50, capped at 200).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "contact_id",
            "in": "query",
            "description": "Only faces labeled with this contact.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "confirmed",
            "in": "query",
            "description": "Filter by exemplar state: `true` = confirmed exemplars only,\n`false` = suggestions only.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "unassigned",
            "in": "query",
            "description": "Only faces with no contact label at all (the review queue).",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of detected faces",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_FaceRefResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/face/refs/{id}/assignment": {
      "put": {
        "tags": [
          "face"
        ],
        "summary": "Set, replace, or clear a face's human label\nPUT /api/face/refs/:id/assignment",
        "operationId": "assign_face_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Face reference ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignFaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Label updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaceRefResponse"
                }
              }
            }
          },
          "400": {
            "description": "Contact does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Face not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssignFaceRequest": {
        "type": "object",
        "description": "Set/replace/clear a face's human label. `null` clears the assignment.",
        "properties": {
          "contact_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
...
      "ContactResponse": {
        "type": "object",
        "description": "A named person faces can be matched to.",
        "required": [
          "id",
          "name",
          "confirmed_faces",
          "total_faces"
        ],
        "properties": {
          "confirmed_faces": {
            "type": "integer",
            "format": "int64",
            "description": "Faces of this contact that are confirmed exemplars."
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "total_faces": {
            "type": "integer",
            "format": "int64",
            "description": "Faces carrying this contact as a (confirmed, pinned, or suggested)\nlabel."
          }
        }
      },
      "CreateContactRequest": {
        "type": "object",
        "description": "Create contact request",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Error response",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "MessageResponse": {
        "type": "object",
        "description": "A generic `{ \"message\": \"...\" }` acknowledgement body.\n\nThe single envelope for endpoints whose only useful response is a\nhuman-readable confirmation (deletes, queued jobs, tag updates, …).\nReplaces the scattered ad-hoc `Json(json!({ \"message\": ... }))` so every\nsuch endpoint shares one typed, OpenAPI-documented shape.",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "UpdateContactRequest": {
        "type": "object",
        "description": "Rename contact request",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}